HarryAndPotter
Two guys Harry and Potter started manipulating one String. Let's say "abcdef".
Harry is removing the last 1 character of the string and append it to start. Handover the string to Potter. - "fabcde"
Potter is removing the last 2 characters of the string and append it to start.  Handover the string to Harry. - "efabcd"
The goal is to get the number of iterations when the resulting string will match the input string.